split discovery into 3 files
